kadai-tasklist-lesson12-9

+++ app/Http/Controllers/TasklistController.php
@@ -94,9 +94,13 @@ class TasklistController extends Controller
     {
         $task = Task::find($id);
         
+        if (\Auth::id() === $task->user_id) {
         return view('tasks.show', [
             'task' => $task,
             ]);
+        }else{
+            return redirect('/');
+        }
     }
 
     /**
@@ -109,9 +113,13 @@ class TasklistController extends Controller
     {
         $task = Task::find($id);
         
+        if (\Auth::id() === $task->user_id) {
         return view('tasks.edit',[
             'task' => $task,
             ]);
+        }else{
+            return redirect('/');
+        }
     }
 
     /**
@@ -145,8 +153,13 @@ class TasklistController extends Controller
     public function destroy($id)
     {
         $task = Task::find($id);
+        
+        if (\Auth::id() === $task->user_id){
         $task->delete();
         
         return redirect('/');
+        }else{
+            return redirect('/');
+        }
     }
 }
